1434 add interface
